has been updated the way how to build and render the elements in the screen, has been added the license plate font and updated the mocks

+++ composeApp/src/commonMain/kotlin/com/cameparkare/dashboardapp/domain/models/terminal/TerminalResponseModel.kt
@@ -4,13 +4,13 @@ import com.cameparkare.dashboardapp.infrastructure.source.remote.dto.TerminalRes
 import com.cameparkare.dashboardapp.infrastructure.source.remote.dto.mapDitsJsonToModel
 
 data class TerminalResponseModel(
-    val dispatcherCode: Int,
+    val dispatcherCode: Long,
     val ditsTUI: List<DitResponseModel>?,
 )
 
 fun TerminalResponseDto.toModel(): TerminalResponseModel {
     return TerminalResponseModel(
-        dispatcherCode = dialog.dialogNumber,
+        dispatcherCode = dialog.dialogNumber.toLong(),
         ditsTUI = mapDitsJsonToModel(ditsTUI).map { it.toModel() }
     )
 }
